refactor: consolidate useFlashAnimation into SDK, remove local copies
