+++ packages/sdk/src/index.ts
@@ -5,6 +5,7 @@ export {PriceProvider, PriceContext} from './providers/PriceProvider';
 export {useAssetPrice, usePrices} from './hooks/usePrices';
 export {useConnectionStatus} from './hooks/useConnectionStatus';
 export {useHistoricalPrices} from './hooks/useHistoricalPrices';
+export {useFlashAnimation} from './hooks/useFlashAnimation';
 export {ASSETS, ASSET_MAP, KRAKEN_PAIR_TO_SYMBOL} from './constants';
 export type {Asset, AssetSymbol, PriceUpdate, PriceMap} from './types';
 export {formatPrice, formatValue} from './utils/formatPrice';
