+++ packages/trading/src/components/AssetRow.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import {Image, Pressable, StyleSheet, Text, View} from 'react-native';
 import Animated from 'react-native-reanimated';
-import {useAssetPrice, formatPrice, getAssetIconUri, colors, type Asset} from 'super-app-showcase-sdk';
-import {useFlashAnimation} from '../hooks/useFlashAnimation';
+import {useAssetPrice, useFlashAnimation, formatPrice, getAssetIconUri, colors, type Asset} from 'super-app-showcase-sdk';
 
 interface AssetRowProps {
   asset: Asset;
